+++ src/component/List_Container/css/List_Container.css
@@ -52,7 +52,7 @@
     right: 10px;
     display: block;
     height: 30px;
-    line-height: 30px;
+    line-height: 28px;
     width: 30px;
     text-align: center;
     background: RGBa(45, 175, 90, 0.4);
